+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 * add: drop-in replacement variant `new RError('FOO')`
 
+### 1.0.2 (2016-12-06)
+
+* fix: chain was missing an error
+
+### 1.0.1 (2016-12-05)
+
+* fix: chain contains undefined
+
 ### 1.0.0 (2016-10-28)
 
 * initial release